data_pipeline/requirements.txt
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
Deprecated==1.2.18
idna==3.10
pycparser==2.22
PyGithub>=1.58
PyJWT==2.10.1
PyNaCl==1.5.0
python-dotenv
PyYAML==6.0.2
requests
typing_extensions==4.13.1
unidiff
urllib3==2.3.0
wrapt==1.17.2
